Refactored CountdownCarousel and added it to Schedule componenet
